piscine-data/Data-Science-2/requirements.txt
numpy
pandas
matplotlib
seaborn
flake8
psycopg2-binary
polars
connectorx
pyarrow